# cmd/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This API Handle Shifts.",
        "title": "Drive Fluency",
        "termsOfService": "https://terminos-y-condiciones",
        "contact": {
            "name": "API Support"
        },
        "license": {},
        "version": "1.0"
    },
    "paths": {
        "/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Inicio de sesión",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/packs": {
            "get": {
                "summary": "Buscar packs",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "summary": "Crear pack",
                "parameters": [
                    {
                        "type": "string",
                        "description": "TOKEN",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/packs/{id}": {
            "get": {
                "summary": "Buscar pack",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "put": {
                "summary": "Modificar pack",
                "parameters": [
                    {
                        "type": "string",
                        "description": "TOKEN",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "summary": "Eliminar pack",
                "parameters": [
                    {
                        "type": "string",
                        "description": "TOKEN",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            },
            "patch": {
                "summary": "Modificar un campo del pack",
                "parameters": [
                    {
                        "type": "string",
                        "description": "TOKEN",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    }
}